Crate browser_window

Browser Window is a Rust crate that allows you to utilize a browser engine to build graphical interfaces. The idea is similar to Electron.

Warning

Browser Window is still in alpha stage of development. The API is bound to have breaking changes every minor version increment. Once version 1.0.0 is released, minor version increments will be backwards-compatible.

Warning

At the moment only Windows is suppored, but Linux support is on the way.

Moreover, the only supported browser engine right now is Chromium's Blink. Therefore, the Chromium Embedding Framework (CEF) is used as a dependency. To get CEF set up properly, enable feature cef, and take a look here.

Getting Started

To start building apps with Browser Window, take a quick look at the application module.

For a rich example, see this example code.

Modules

application
This module contains runtime and application related handles.

browser
This module contains all browser related handles and stuff.

event
This module contains all event related types.

prelude
Some common traits that need to be often available.

window
This module contains all window related functionality.

Structs

DelegateFuture
This future executes a closure on the GUI thread and returns the result.

DelegateFutureFuture
This future runs a future on the GUI thread and returns its output.

Enums

DelegateError
The error that occurs when you're delegating work to the GUI thread, but it fails to finish and/or return a result.

Traits

Property
A property is something that has a setter and a getter.