Module browser_window::browser

This module contains all browser related handles and stuff.

Keep in mind that BrowserWindow exposes the same methods as BrowserWindowHandle and WindowHandle does. The methods of BrowserWindowHandle are displayed correctly at the page of BrowserWindow, but the methods of WindowHandle are not displayed. Be sure to check them out here.

Structs

BrowserWindow
An owned browser window handle. When this handle goes out of scope, its resource get scheduled for cleanup.

BrowserWindowBuilder
Used to create a BrowserWindow or BrowserWindowThreaded instance, depending on whether or not you have feature threadsafe enabled.

BrowserWindowHandle
This is a handle to an existing browser window.

BrowserWindowThreaded
Note: Only available with feature threadsafe enabled.

Enums

Source
The type of content to display in a browser window

Traits

OwnedBrowserWindow

Type Definitions

BrowserDelegateFuture
The future that dispatches a closure on the GUI thread.