Struct browser_window::browser::BrowserWindowBuilder

pub struct BrowserWindowBuilder { /* fields omitted */ }

Used to create a BrowserWindow or BrowserWindowThreaded instance, depending on whether or not you have feature threadsafe enabled.

Warning

BrowserWindowBuilder dereferences to WindowBuilder, so that you can modify anything related to the window as well. This comes with a catch though. If you call any functions from WindowBuilder, they return a reference to WindowBuilder. So the following doesn't work:

let bwb = BrowserWindowBuilder::new( Source::Url("https://www.duckduckgo.com".into()) )
	.title("DuckDuckGo")
	.dev_tools(true);

This is because the call to title returns a reference to WindowBuilder, which doesn't have a method dev_tools.

The following is also wrong:

let bw = BrowserWindowBuilder::new( Source::Url("https://www.duckduckgo.com".into()) )
	.dev_tools(true)
	.title("DuckDuckGo")
	.build( app );

This is because Browser Window currently does not support creating windows that don't have browsers in them. build is a method of BrowserWindowBuilder, yet title returns a reference to WindowBuilder, which has no build method.

The solution is to do this:

let mut bwb = BrowserWindowBuilder::new( Source::Url("https://www.duckduckgo.com".into()) )
	.dev_tools(true);
bwb.title("DuckDuckGo");
let bw = bwb.build( app );

Implementations

impl BrowserWindowBuilder

pub fn async_handler<H, F>(&mut self, handler: H) -> &mut Self where
H: FnMut(BrowserWindowHandle, String, Vec<String>) -> F + Send + 'static,
F: Future<Output = ()> + 'static,

Configure a closure that can be invoked from within JavaScript. The closure's second parameter specifies a command name. The closure's third parameter specifies an array of string arguments.

pub fn dev_tools(&mut self, enabled: bool) -> &mut Self

Sets whether or not an extra window with developer tools will be opened together with this browser. When in debug mode the default is true. When in release mode the default is false.

pub fn new(source: Source) -> Self

Creates an instance of a browser window builder.

Arguments

source - The content that will be displayed in the browser window.

pub async fn build(
self,
app: ApplicationHandleThreaded
) -> Result<BrowserWindowThreaded, DelegateError>

Creates the browser window.

Keep in mind that the description of this function is for when feature threadsafe is enabled. When it is not enabled, it looks like this:

pub async fn build( self, app: ApplicationHandle ) -> BrowserWindow { /* ... */ }

Arguments

app - An (thread-safe) application handle.