Struct browser_window::browser::BrowserWindowHandle

pub struct BrowserWindowHandle { /* fields omitted */ }

This is a handle to an existing browser window.

Implementations

impl BrowserWindowHandle

pub fn app(&self) -> ApplicationHandle

Returns the application handle associated with this browser window.

pub async fn eval_js<'_, '_>(
&'_ self,
js: &'_ str
) -> Result<String, JsEvaluationError>

Executes the given javascript code and returns the output as a string. If you don't need the result, see exec_js.

pub fn exec_js(&self, js: &str)

Executes the given javascript code without waiting on it to finish.

pub fn navigate(&self, url: &str)

Causes the browser to navigate to the given url.

pub fn window(&self) -> WindowHandle