Module browser_window::window

This module contains all window related functionality.

Structs

ContentDimensions
Gets or sets the width and height of the content of the window.

Opacity
Gets or sets the opacity of the window. An opacity of 255 means the window is invisible. An opacity of 0 means the window is completely visible. Anything in between makes the window transparent.

Position
Gets or sets the current position of the window.

Title
Gets or sets the title of the window.

WindowBuilder
Exposes functionality related to constructing a window.

WindowDimensions
Gets or sets the current window size including its border and titlebar.

WindowHandle
A handle that exposes all windowing functionality.

WindowResizeEventArgs

Traits

OwnedWindow

Type Definitions

StandardWindowEvent